fix : fixing clang-tidy errors

+++ numerical_methods/fast_fourier_transform.cpp
@@ -38,7 +38,7 @@ namespace numerical_methods {
  * @returns y if n!=1
  */
 
-std::complex<double>* FastFourierTransform(std::complex<double>*p,double n)
+std::complex<double>* FastFourierTransform(std::complex<double>*p,unsigned long int n)
 {
 
 	if(n==1){
@@ -106,8 +106,8 @@ std::complex<double>* FastFourierTransform(std::complex<double>*p,double n)
     std::complex<double> t1[2]={1,2}; /// Test case 1
     std::complex<double> t2[4]={1,2,3,4}; /// Test case 2
 
-    double n1 = 2;
-    double n2 = 4;
+    unsigned long int n1 = 2;
+    unsigned long int n2 = 4;
     std::vector<std::complex<double>> r1 = {
         {3, 0}, {-1, 0}};  /// True Answer for test case 1
